Backend/mt5-sync-service/requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.13
